+++ packages/core-js-compat/src/modules-by-versions.js
@@ -80,4 +80,7 @@ module.exports = {
     'esnext.typed-array.find-last-index',
     'esnext.typed-array.unique-by',
   ],
+  3.11: [
+    'esnext.object.has-own',
+  ],
 };