+++ packages/core-js/internals/has.js
@@ -1,5 +1,5 @@
 var hasOwnProperty = {}.hasOwnProperty;
 
-module.exports = function (it, key) {
+module.exports = function hasOwn(it, key) {
   return hasOwnProperty.call(it, key);
 };