+++ packages/core-js/stage/2.js
@@ -1,3 +1,4 @@
+require('../proposals/accessible-object-hasownproperty');
 require('../proposals/array-find-from-last');
 require('../proposals/array-is-template-object');
 require('../proposals/iterator-helpers');