+++ tests/compat/tests.js
@@ -1320,6 +1320,9 @@ GLOBAL.tests = {
   'esnext.number.range': function () {
     return Number.range;
   },
+  'esnext.object.has-own': function () {
+    return Object.hasOwn;
+  },
   // TODO: Remove from `core-js@4`
   'esnext.object.iterate-entries': function () {
     return Object.iterateEntries;